# Mammoth_Labs/blockscan-api/internal/adapter/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Backend Template",
        "title": "blockscan-go API",
        "contact": {},
        "license": {
            "name": "Go Echo"
        },
        "version": "1.0"
    },
    "host": "tracking.gmmtchain.io",
    "basePath": "/api/v1",
    "paths": {
...
        "/giant_mammoth/staking/info/{chain_id}/{wallet_address}": {
            "get": {
                "description": "Get Staking By Account / 해당 지갑에 관련된 staking 정보를 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GiantMammoth"
                ],
                "summary": "Get Staking By Account / *공용",
                "operationId": "GetStakingByAccount",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "지갑주소",
                        "name": "wallet_address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_blockchain_gmmt_staking_ValidatorByAccountData"
                        }
                    }
                }
            }
        },
        "/giant_mammoth/staking/list/{chain_id}": {
            "get": {
                "description": "Get Staking list / Gmmt staking 정보를 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GiantMammoth"
                ],
                "summary": "Get Staking list / *공용",
                "operationId": "GetStakingList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_blockchain_gmmt_staking_Staking"
                        }
                    }
                }
            }
        },
        "/giant_mammoth/swap/list/{chain_id}": {
            "get": {
                "description": "Get Swap Pair List / 아이보리 스왑 pair list를 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GiantMammoth"
                ],
                "summary": "Get Swap Pair List / *공용",
                "operationId": "GetSwapPairList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_blockchain_gmmt_swap_Swap"
                        }
                    }
                }
            }
        },
        "/log/nft/activities/{wallet_address}": {
            "get": {
                "description": "GetByWallet / 해당 지갑에서 발생한 NFT 로그 기록 가져오는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Log"
                ],
                "summary": "Get NFTs by wallet / *공용",
                "operationId": "LogGetNFTsByWallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "월렛주소",
                        "name": "wallet_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "가져올 아이템 수 최대 100",
                        "name": "take",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "건너뛸 아이템 수 (옵션) 기본 0",
                        "name": "skip",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_log_GetLogNFTsByWalletData"
                        }
                    }
                }
            },
            "post": {
                "description": "LogGetSearchNFTsByWallet / 해당 지갑에서 발생한 NFT 로그 날짜, 종류에 따라 가져오는 API / *공용(gmwallet, edem)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Log"
                ],
                "summary": "Get NFT logs by wallet / *공용",
                "operationId": "LogGetSearchNFTsByWallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "월렛주소",
                        "name": "wallet_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_core_domain_log.GetSearchByWalletInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_log_GetLogNFTsByWalletData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_blockchain_gmmt_staking_ValidatorByAccountData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_blockchain_gmmt_staking.ValidatorByAccountData"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_blockchain_gmmt_swap_Swap": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_blockchain_gmmt_swap.Swap"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
...
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_log_GetLogNFTsByWalletData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_log.GetLogNFTsByWalletData"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
...
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_blockchain_gmmt_staking_Staking": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/blockscan-go_internal_core_blockchain_gmmt_staking.Staking"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
...
        "blockscan-go_internal_core_blockchain_gmmt_staking.Staking": {
            "type": "object",
            "properties": {
                "apr": {
                    "type": "string"
                },
                "bonded_coin": {
                    "type": "string"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_blockchain_gmmt_staking.Validator"
                    }
                }
            }
        },
        "blockscan-go_internal_core_blockchain_gmmt_staking.Validator": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "changed_at": {
                    "type": "integer"
                },
                "claimed_at": {
                    "type": "integer"
                },
                "commission_rate": {
                    "type": "integer"
                },
                "jailed_before": {
                    "type": "integer"
                },
                "slashes_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "total_delegated": {
                    "type": "string"
                },
                "total_rewards": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_blockchain_gmmt_staking.ValidatorByAccountData": {
            "type": "object",
            "properties": {
                "claimable_asset": {
                    "type": "string"
                },
                "commission_rate": {
                    "type": "integer"
                },
                "delegation": {
                    "type": "string"
                },
                "jailed_before": {
                    "type": "integer"
                },
                "owner_address": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "pending_asset": {
                    "type": "string"
                },
                "slashes_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "total_delegated": {
                    "type": "string"
                },
                "validator_address": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "blockscan-go_internal_core_blockchain_gmmt_swap.Swap": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer"
                },
                "liquidity": {
                    "type": "string"
                },
                "pair_address": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "size_up": {
                    "type": "boolean"
                },
                "token_0_address": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "token_0_decimals": {
                    "type": "integer"
                },
                "token_0_get_reserves": {
                    "type": "string"
                },
                "token_0_name": {
                    "type": "string"
                },
                "token_0_symbol": {
                    "type": "string"
                },
                "token_1_address": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "token_1_decimals": {
                    "type": "integer"
                },
                "token_1_get_reserves": {
                    "type": "string"
                },
                "token_1_name": {
                    "type": "string"
                },
                "token_1_symbol": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_common_utils.ErrorType": {
            "type": "string",
            "enum": [
                "No data found",
                "Syntax",
                "Duplicate",
                "Invalid sql statement",
                "Bad connection",
                "Invalid character",
                "Foreign key",
                "Check violation",
                "Success",
                "Fail",
                "Not found",
                "Invalid input",
                "Database",
                "Internal"
            ],
            "x-enum-varnames": [
                "NoDataFound",
                "SyntaxError",
                "DuplicateError",
                "InvalidSQLStatementError",
                "BadConnectionError",
                "JsonParseError",
                "ForeignKeyError",
                "CheckViolationError",
                "Success",
                "Fail",
                "NotFound",
                "InvalidInput",
                "DatabaseError",
                "InternalError"
            ]
        },
...
        "blockscan-go_internal_core_domain_log.GetLogNFTsByWalletData": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "block_hash": {
                    "type": "string"
                },
                "block_number": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "dapp": {
                    "type": "string"
                },
                "erc1155_token_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "erc1155_url": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "erc1155_value": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "from_address": {
                    "type": "string"
                },
                "function": {
                    "type": "string"
                },
                "gas": {
                    "type": "string"
                },
                "gas_price": {
                    "type": "string"
                },
                "gas_used": {
                    "type": "string"
                },
                "market_info_array": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_log.MarketInfo"
                    }
                },
                "name": {
                    "type": "string"
                },
                "nft_volume_array": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_log.NftVolumeData"
                    }
                },
                "symbol": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "to_address": {
                    "type": "string"
                },
                "token_id": {
                    "type": "string"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "trade_nft_volume": {
                    "type": "string"
                },
                "trade_nft_volume_contract": {
                    "type": "string"
                },
                "trade_nft_volume_symbol": {
                    "type": "string"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_domain_log.GetSearchByWalletInput": {
            "type": "object",
            "required": [
                "chain_id",
                "date",
                "take",
                "type",
                "walletAddress"
            ],
            "properties": {
                "chain_id": {
                    "type": "integer"
                },
                "date": {
                    "type": "string",
                    "enum": [
                        "all",
                        "7D",
                        "30D",
                        "1Y"
                    ]
                },
                "skip": {
                    "type": "integer"
                },
                "take": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "all",
                        "mint",
                        "transfer",
                        "sale",
                        "burn"
                    ]
                },
                "walletAddress": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_domain_log.MarketInfo": {
            "type": "object",
            "properties": {
                "buyer": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "collection": {
                    "type": "string"
                },
                "seller": {
                    "type": "string"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "volume": {
                    "type": "string"
                },
                "volume_contract_address": {
                    "type": "string"
                },
                "volume_symbol": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_domain_log.NftVolumeData": {
            "type": "object",
            "properties": {
                "chain_id": {
                    "type": "integer"
                },
                "contract": {
                    "type": "string"
                },
                "event": {
                    "type": "string"
                },
                "from_address": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "to_address": {
                    "type": "string"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    }
}